# lib/api-zod/src/generated/types/extractionResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stablecoin Research Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractionResult {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  abstract?: string | null;
  keywords?: string[];
  authors?: string[];
  /** @nullable */
  published_date?: string | null;
  /** @nullable */
  journal?: string | null;
  /** @nullable */
  doi?: string | null;
}