# lib/api-zod/src/generated/types/regulatoryEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stablecoin Research Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface RegulatoryEntry {
  id: number;
  country: string;
  /** @nullable */
  region?: string | null;
  /** @nullable */
  authority?: string | null;
  title: string;
  /** @nullable */
  title_zh?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  summary_zh?: string | null;
  /** @nullable */
  document_url?: string | null;
  effective_date: string;
  /** @nullable */
  category?: string | null;
  created_at: string;
}