# lib/api-zod/src/generated/types/regulatoryTimelineYear.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stablecoin Research Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { RegulatoryEntry } from './regulatoryEntry';

export interface RegulatoryTimelineYear {
  year: number;
  entries: RegulatoryEntry[];
}