# lib/api-zod/src/generated/types/researchPaper.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stablecoin Research Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface ResearchPaper {
  id: number;
  title: string;
  /** @nullable */
  title_zh?: string | null;
  /** @nullable */
  abstract?: string | null;
  /** @nullable */
  abstract_zh?: string | null;
  key_innovations?: string[];
  key_innovations_zh?: string[];
  authors?: string[];
  /** @nullable */
  published_date?: string | null;
  /** @nullable */
  pdf_url?: string | null;
  keywords?: string[];
  created_at: string;
}